+++ src/fields/Link.php
@@ -512,22 +512,10 @@ class Link extends Field implements InlineEditableFieldInterface, RelationalFiel
             $value instanceof LinkData &&
             $element?->propagating &&
             ($element->propagateAll || ($element->isNewForSite && !isset($element->duplicateOf))) &&
-            isset($element->propagatingFrom)
+            isset($element->propagatingFrom) &&
+            $this->getTranslationKey($element) !== $this->getTranslationKey($element->propagatingFrom)
         ) {
-            // in order to avoid infinite loop when using custom translation format with a translation key containing `include()`
-            // we need to prevent `View::renderObjectTemplate()` from trying to normalize this value again and again
-            // to do that, we can e.g. set `propagating` to false before getting the translation key
-            // see https://github.com/craftcms/cms/issues/18363 for more details
-            if ($this->translationMethod === self::TRANSLATION_METHOD_CUSTOM) {
-                $element->propagating = false;
-            }
-
-            if ($this->getTranslationKey($element) !== $this->getTranslationKey($element->propagatingFrom)) {
-                $value = $this->localizeLinkValue($value, $element);
-            }
-
-            // set $propagating back to true
-            $element->propagating = true;
+            $value = $this->localizeLinkValue($value, $element);
         }
 
         // as above but specifically for nested entries when field uses "none" propagation method